💄 MyProductCard 컴포넌트 생성, ProductTable에 적용
